[sandbox][string] Add String::Utf8Length and String::EncodeUtf8
Bug: 373485796, 376071292

+++ src/codegen/compiler.cc
@@ -3709,7 +3709,7 @@ class StressBackgroundCompileThread : public ParkingThread {
     }
 
    private:
-    uint32_t source_length_;
+    size_t source_length_;
     std::unique_ptr<char[]> source_buffer_;
     bool done_;
   };
